### 2025-06-04 3df7837
add necessary docs: code of conduct, security, citation
