+++ README.md
@@ -54,28 +54,25 @@ This initiative is designed to streamline your development process, ensuring tha
 
 Furthermore, we recognize the importance of proper usage and attribution of the utility functions we have integrated into pyufunc. If any package developer finds their utility function has been used improperly, we encourage you to reach out to the pyufunc developers for further discussion. We are committed to maintaining a respectful and collaborative relationship with the original developers, ensuring that all contributions are appropriately acknowledged and utilized within the bounds of open-source licenses and community norms. Your feedback and insights are crucial to us, as they help in refining pyufunc to better serve the open-source community.
 
-
-|     Package Name     | Latest Version                                                                                              | Latest Commit                                                                                                                                                  | Repository                                                                                              | Description                                                                                                                                                                                                            |
-| :------------------: | ----------------------------------------------------------------------------------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------------------- | ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-|      [@pyutil]       | [![PyPI version](https://badge.fury.io/py/pyutil.svg)](https://badge.fury.io/py/pyutil)                     | [![img](https://img.shields.io/github/last-commit/tpltnt/pyutil.svg)](https://img.shields.io/github/last-commit/tpltnt/pyutil)                                 | [GitHub Repo 1](https://github.com/tpltnt/pyutil)<br />[GitHub Repo 2](https://github.com/zooko/pyutil) | a library of useful Python functions and classes                                                                                                                                                                       |
-| [@Fu_PyHelpers_2020] | [![PyPI version](https://badge.fury.io/py/pyhelpers.svg)](https://badge.fury.io/py/pyhelpers)               | [![img](https://img.shields.io/github/last-commit/mikeqfu/pyhelpers.svg)](https://img.shields.io/github/last-commit/mikeqfu/pyhelpers)                         | [GitHub Repo](https://github.com/mikeqfu/pyhelpers)                                                     | An open-source toolkit for facilitating Python<br />users' data manipulation tasks                                                                                                                                     |
-|      [@psutil]       | [![PyPI version](https://badge.fury.io/py/psutil.svg)](https://badge.fury.io/py/psutil)                     | [![img](https://img.shields.io/github/last-commit/giampaolo/psutil.svg)](https://img.shields.io/github/last-commit/giampaolo/psutil)                           | [GitHub Repo](https://github.com/giampaolo/psutil)                                                      | Cross-platform lib for process and system monitoring in Python                                                                                                                                                         |
-|    [@pyutilator]     | [![PyPI version](https://badge.fury.io/py/pyutilator.svg)](https://badge.fury.io/py/pyutilator)             | [![img](https://img.shields.io/github/last-commit/antoprince001/pyutilator.svg)](https://img.shields.io/github/last-commit/antoprince001/pyutilator)           | [GitHub Repo](https://github.com/antoprince001/pyutilator)                                              | open source python package comprising of decorators<br /> that can be used for utility operations                                                                                                                      |
[…]
