### 2025-06-04 #8344 @arfon
Hi @xyluo25, I noticed you've been soliciting potential reviewers for your submission. e.g.,

- https://github.com/openjournals/joss-reviews/issues/3274#issuecomment-2941579266
- https://github.com/openjournals/joss-reviews/issues/2214#issuecomment-2941582216
- https://github.com/openjournals/joss-reviews/issues/3607#issuecomment-2941597986
- https://github.com/openjournals/joss-reviews/issues/5076#issuecomment-2941567911

While I appreciate the proactive nature of your efforts, finding reviewers is something that the assigned editor will (and should) handle.

/ cc @crvernon
