### 2025-06-15 6cfdb55
add func: find_executable_from_PATH_on_linux
