+++ pyufunc/util_pathio/_path.py
@@ -11,6 +11,7 @@ from pathlib import Path
 import os
 import re
 import sys
+import subprocess
 from typing import Callable, Union
 from pyufunc.pkg_configs import config_color
 
@@ -646,6 +647,90 @@ def find_executable_from_PATH_on_win(exe_name: str,
     Location:
         pyufunc/util_pathio/_path.py
 
+    Note:
+        This function is same as `find_fname_from_PATH_on_win`, both are used to find the executable from the system PATH.
+
+    Examples:
+        >>> import pyufunc as pf
+        >>> pf.find_executable_from_PATH_on_win('python', 'exe', True)
+        ["**/python.exe", ...]
+
+    Returns:
+        list or None: A list of full path of the executable if found, otherwise None.
+    """
+
+    # check if the executable name is a string
+    if not isinstance(exe_name, str):
+        raise ValueError("exe_name should be a string.")
+
+    # check if extension is str
+    if not isinstance(ext, str):
+        raise ValueError("ext should be a string.")
+
+    # check if sel_dir is a list
+    if not isinstance(sel_dir, (list, type(None))):
+        raise ValueError("sel_dir should be a list.")
+
+    # add the selected directories to the system PATH
+    if sel_dir:
+        for dir_ in sel_dir:
+            if os.path.isdir(dir_) and dir_ not in os.getenv("PATH"):
+                os.environ["PATH"] += os.pathsep + dir_
+            elif verbose:
+                print(f"  :The dir: {dir_} is not a valid directory "
+                      "or already in the system PATH. Skipped.")
+
+    # check if exe_name has the extension
+    file_name, ext_str = os.path.splitext(exe_name)
+    if not ext_str:
+        if verbose:
+            print(f"  :The executable: {exe_name} has no extension. Added {ext} as the extension.")
+        exe_name = f"{exe_name}.{ext}"
+
+    # get the full environment PATH
+    env_paths = os.getenv("PATH").split(os.pathsep)
+
+    res = []
+    for path in env_paths:
+        abs_path = path2linux(os.path.join(path, exe_name))
+
+        # check if the file exists and is executable
+        if os.path.isfile(exe_name) and os.access(abs_path, os.X_OK):
+            res.append(abs_path)
+
+    if not res:
+        if verbose:
+            print(f"  :Could not find {exe_name} in the system PATH."
+                  " Please make sure the executable is installed."
+                  " please include executable extension in the name.")
+        return None
+
+    if verbose:
+        print(f"  :Found {exe_name} in the system PATH:")
+        for path in res:
+            print(f"  :  {path}")
+    return res
+
+
+def find_executable_on_win(exe_name: str,
+                                     ext: str = "exe",
+                                     *,
+                                     sel_dir: list = None,
+                                     verbose: bool = True) -> list | None:
+    """Find the executable from the system PATH.
+
+    Args:
+        exe_name (str): The executable name to search for.
+        ext (str): The extension of the executable. Defaults to "exe" for executable files.
+        sel_dir (list): The selected directories to search for the executable. Defaults to [].
+        verbose (bool): Whether to print the process info. Defaults to True.
+
+    Location:
+        pyufunc/util_pathio/_path.py
+
+    Note:
+        This function is same as `find_fname_from_PATH_on_win`, both are used to find the executable from the system PATH.
+
     Examples:
         >>> import pyufunc as pf
         >>> pf.find_executable_from_PATH_on_win('python', 'exe', True)
@@ -724,6 +809,9 @@ def find_fname_from_PATH_on_win(fname: str,
     Location:
         pyufunc/util_pathio/_path.py
 
+    Note:
+        This function is same as `find_executable_from_PATH_on_win`, both are used to find the filename from the system PATH.
+
     Examples:
[…]
