### 2025-06-15 e62295f
enhanced gmns_read_link print out message
