### 2025-06-15 a7e5541
update gmns read functions
