### 2025-07-25 #8344 @juliaferraioli
👋🏼 @ jonjoncardoso, @sunnypatel165, @aslan-ng, & @shrishar, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these [guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html). Please let me know even if you're not interested, so that I can start looking for other reviewers.

For a couple of you, it would be your first review, and this is a great way to jump in!
