### 2025-07-29 1ea933a
add project dependencies
