+++ .github/workflows/github_dependents_info.yml
@@ -62,7 +62,7 @@ jobs:
           commit-message: "[GitHub Dependents Info] Updated markdown file(s)"
           delete-branch: true
           title: "[GitHub Dependents Info] Updated markdown file"
-          body: "_Generated with [github-dependents-info](https://github.com/nvuillam/github-dependents-info), by [Nicolas Vuillamy](https://github.com/nvuillam)_"
+          body: "Update markdown file with the latest dependents information."
           labels: documentation
       - name: Create PR output
         run: |
