+++ README.md
@@ -1,4 +1,5 @@
-[![PyPI version](https://badge.fury.io/py/pyufunc.svg)](https://badge.fury.io/py/pyufunc)[![Downloads](https://static.pepy.tech/badge/pyufunc)](https://pepy.tech/project/pyufunc)[![](https://img.shields.io/pypi/wheel/gensim.svg)](https://pypi.org/project/pyufunc/)[![](https://img.shields.io/pypi/pyversions/pyufunc.svg)](https://www.python.org/)[![](https://img.shields.io/github/release-date/xyluo25/pyufunc.svg)](https://img.shields.io/github/release-date/xyluo25/pyufunc.svg)[![](https://readthedocs.org/projects/pyufunc/badge/?version=latest)](https://pyufunc.readthedocs.io/en/latest/?badge=latest)[![](https://img.shields.io/badge/License-MIT-blue.svg)](https://opensource.org/license/mit)<!-- gh-dependents-info-used-by-start --><!-- gh-dependents-info-used-by-end -->
+[![PyPI version](https://badge.fury.io/py/pyufunc.svg)](https://badge.fury.io/py/pyufunc)[![Downloads](https://static.pepy.tech/badge/pyufunc)](https://pepy.tech/project/pyufunc)[![](https://img.shields.io/pypi/wheel/gensim.svg)](https://pypi.org/project/pyufunc/)[![](https://img.shields.io/pypi/pyversions/pyufunc.svg)](https://www.python.org/)[![](https://img.shields.io/github/release-date/xyluo25/pyufunc.svg)](https://img.shields.io/github/release-date/xyluo25/pyufunc.svg)[![](https://readthedocs.org/projects/pyufunc/badge/?version=latest)](https://pyufunc.readthedocs.io/en/latest/?badge=latest)[![](https://img.shields.io/badge/License-MIT-blue.svg)](https://opensource.org/license/mit)<!-- gh-dependents-info-used-by-start -->
+[![Generated by github-dependents-info](https://img.shields.io/static/v1?label=Used%20by&message=13&color=informational&logo=slickpic)](https://github.com/xyluo25/pyufunc/blob/main/docs/static/github-dependents-info.md)<!-- gh-dependents-info-used-by-end -->
 
 # pyufunc
 
