### 2025-07-29 90de6c9
Update pkg_configs.py
