### 2025-08-04 #8746 @editorialbot
[REVIEW]: pyufunc: A Set of Utility Functions that Keep Python Sweet
**Submitting author:** @xyluo25 (<a href="http://orcid.org/0009-0003-1290-9983">Xiangyong Luo</a>)
**Repository:** https://github.com/xyluo25/pyufunc
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.4.1
