### 2025-08-04 #8746 @juliaferraioli
👋🏼 @xyluo25, @shrishar, @jonjoncardoso, @amcandio this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```
