### 2025-08-10 xyluo25/pyufunc#3 @amcandio
JOSS review: incomplete modules
There are some modules that are empty and show up in the docs, for example https://pyufunc.readthedocs.io/en/latest/reference/util_optimization.html. This can be confusing to users

https://github.com/openjournals/joss-reviews/issues/8746
