### 2025-08-10 xyluo25/pyufunc#4 @amcandio
JOSS review: pandas dependency vs auto-installed dependencies
In the README it says:

```
🚀️ No dependencies will be installed to your coding environment unless use of functions that need certain dependency
```

But in requirement it says:
```
pandas>=2.2.0
```

https://github.com/openjournals/joss-reviews/issues/8746
