### 2025-08-10 xyluo25/pyufunc#5 @amcandio
JOSS review: Make auto-install optional
Library currently auto-installs missing dependencies when a function is called. This can be problematic in CI or in offline environments. It can be considered a security concern to make a library autoinstall packages without users knowing.

Consider making feature optional and disabled by default.

https://github.com/openjournals/joss-reviews/issues/8746
