### 2025-08-10 xyluo25/pyufunc#7 @amcandio
JOSS review: licensing concerns
There is an explicit mention in the README that package extracts functionality from other libraries. It is unclear to me what parts of the library were taken from which library, and under which licence. Depending on the other libraries licenses, pyufunc license might be in violation.

There is a sentence in the README to "Your contributions have significantly enriched our open-source community. If any code violates your licensing agreements, please don't hesitate to [Contact Developers](mailto:luoxiangyong01@gmail.com)." However, this is does not prevent any license violation.
