### 2025-08-12 xyluo25/pyufunc#3 @xyluo25
Hi @amcandio, thank you for the feedback on the empty modules. The original plan was to include these categories as placeholders and add functions over time. I understand this can be confusing for users, so I’ll remove the empty modules from the docs (new release). Will keep you updated once it's done. Thanks again.
