### 2025-08-12 xyluo25/pyufunc#5 @xyluo25
Hi @amcandio, great suggestion. Our original design aimed to minimize install-time dependencies at the project PI’s request. I’ll add a global configuration flag, disabled by default, to control this behavior. Thanks again for the useful feedback.
