### 2025-08-12 #8746 @xyluo25
> [@juliaferraioli](https://github.com/juliaferraioli) I have some concerns with respect licensing of this package ([xyluo25/pyufunc#7](https://github.com/xyluo25/pyufunc/issues/7)).
>
> There is an explicit mention in the README that package extracts functionality from other libraries. It is unclear to me what parts of the library were taken from which library, and under which licence. From what I see on the mentioned packages, there is a mix between Apache-2.0, MIT, BSD-3 and also unlicensed.
>
> Depending on the other libraries licenses, pyufunc license might be in violation.
>
> How should we proceed here?

Hi @juliaferraioli and @amcandio, could you please confirm whether my response here (https://github.com/xyluo25/pyufunc/issues/7#issuecomment-3179552829) will addresses the licensing requirements? I am happy to make any additional changes. thanks.
