### 2025-08-13 #8746 @juliaferraioli
> [@juliaferraioli](https://github.com/juliaferraioli) where should I submit my review? Should I tick the checklist?

Yes! You can review the [criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) as you do so. You may submit issues/pull requests directly to the repository or raise them here. We just ask that you mention openjournals/joss-reviews#8746 if you submit issues/PRs on the repo so we have visibility of them here.
