+++ docs/static/github-dependents-info.md
@@ -3,22 +3,22 @@
 [![Generated by github-dependents-info](https://img.shields.io/static/v1?label=Used%20by&message=13&color=informational&logo=slickpic)](https://github.com/xyluo25/pyufunc/network/dependents)
 [![Generated by github-dependents-info](https://img.shields.io/static/v1?label=Used%20by%20(public)&message=13&color=informational&logo=slickpic)](https://github.com/xyluo25/pyufunc/network/dependents)
 [![Generated by github-dependents-info](https://img.shields.io/static/v1?label=Used%20by%20(private)&message=-13&color=informational&logo=slickpic)](https://github.com/xyluo25/pyufunc/network/dependents)
-[![Generated by github-dependents-info](https://img.shields.io/static/v1?label=Used%20by%20(stars)&message=108&color=informational&logo=slickpic)](https://github.com/xyluo25/pyufunc/network/dependents)
+[![Generated by github-dependents-info](https://img.shields.io/static/v1?label=Used%20by%20(stars)&message=111&color=informational&logo=slickpic)](https://github.com/xyluo25/pyufunc/network/dependents)
 
 | Repository | Stars  |
 | :--------  | -----: |
 |<img class="avatar mr-2" src="https://avatars.githubusercontent.com/u/74271152?s=40&v=4" width="20" height="20" alt="">  &nbsp; [asu-trans-ai-lab](https://github.com/asu-trans-ai-lab) / [grid2demand](https://github.com/asu-trans-ai-lab/grid2demand) | 84 |
-|<img class="avatar mr-2" src="https://avatars.githubusercontent.com/u/119607574?s=40&v=4" width="20" height="20" alt="">  &nbsp; [ORNL-Real-Sim](https://github.com/ORNL-Real-Sim) / [Real-Twin](https://github.com/ORNL-Real-Sim/Real-Twin) | 9 |
-|<img class="avatar mr-2" src="https://avatars.githubusercontent.com/u/11805788?s=40&v=4" width="20" height="20" alt="">  &nbsp; [alphavector](https://github.com/alphavector) / [all](https://github.com/alphavector/all) | 4 |
+|<img class="avatar mr-2" src="https://avatars.githubusercontent.com/u/119607574?s=40&v=4" width="20" height="20" alt="">  &nbsp; [ORNL-Real-Sim](https://github.com/ORNL-Real-Sim) / [Real-Twin](https://github.com/ORNL-Real-Sim/Real-Twin) | 10 |
+|<img class="avatar mr-2" src="https://avatars.githubusercontent.com/u/11805788?s=40&v=4" width="20" height="20" alt="">  &nbsp; [alphavector](https://github.com/alphavector) / [all](https://github.com/alphavector/all) | 5 |
 |<img class="avatar mr-2" src="https://avatars.githubusercontent.com/u/36498464?s=40&v=4" width="20" height="20" alt="">  &nbsp; [xyluo25](https://github.com/xyluo25) / [grid2demand](https://github.com/xyluo25/grid2demand) | 3 |
 |<img class="avatar mr-2" src="https://avatars.githubusercontent.com/u/25952589?s=40&v=4" width="20" height="20" alt="">  &nbsp; [Lyn4ever29](https://github.com/Lyn4ever29) / [pipy_server](https://github.com/Lyn4ever29/pipy_server) | 3 |
[…]
