### 2025-09-03 #8746 @xyluo25
> Hi [@xyluo25](https://github.com/xyluo25), I've been doing more thinking around your submission, and I think the right path forward is to pause it until we've had the chance to run it by the editorial board to ensure it is in scope. You can see my earlier comment as to why I'm recommending this. I apologize for the delay in following up on this, but I have been thinking about it quite a bit from a number of different angles!

Hi @juliaferraioli, thank you for the thoughtful follow up. I agree with pausing the submission and I appreciate the care you are taking. I am ready to tighten the core functionality, streamline the API, and strengthen documentation and examples to align with expectations, especially comments from @amcandio. Let me move forward with this so the tool can become more useful to a wider range of developers!
